repr: hoist strconv tests into integration test file
These are testing public functionality, so they should go into an
integration test file as a matter of philosophy.
